requirements.txt
# Framework Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.0.3

# Visão Computacional e IA
ultralytics>=8.0.225
opencv-python>=4.8.0
numpy>=1.26.0
Pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0

# Text-to-Speech
gtts==2.4.0
pygame==2.5.2

# Utilitários
python-dotenv==1.0.0
aiofiles==23.2.1

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2